pyproject.toml
[project]
name = "czech-banking-ci"
version = "0.1.0"
description = "Czech Banking Competitive Intelligence Monitor"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "click>=8.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "tenacity>=8.2",
    "python-dateutil>=2.9",
    "openpyxl>=3.1",
    "feedparser>=6.0",
    "flask>=3.0",
    "flask-cors>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
ci-monitor = "src.cli:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]
